server_button/requirements.txt
flask==2.0.1
PyYAML==5.4.1
pyserial==3.5
RPi.GPIO==0.7.1
timeloop==1.0.2
paho-mqtt ==1.6.1
bson ==0.5.10
# Dev requirements
black ==22.3.0
pylint ==2.14.3
